// packages/hex/src/models/globalInfo.ts
export interface GlobalInfo {
	lockedHeartsTotal: bigint;
	nextStakeSharesTotal: bigint;
	shareRate: bigint;
	stakePenaltyTotal: bigint;
	currentDay: bigint;
	stakeSharesTotal: bigint;
	latestStakeId: bigint;
	unclaimedSatoshisTotal: bigint;
	claimedSatoshisTotal: bigint;
	claimedBtcAddrCount: bigint;
	blockTimestamp: bigint;
	totalSupply: bigint;
	transferLobby: bigint;
}